+++ src/DragInput.js
@@ -90,9 +90,11 @@ export default class DragInput{
             case 'touchend':
             case 'touchcancel': // end
                 //if ( (event.type === 'touchend' || event.type === 'touchcancel') && event.cancelable) event.preventDefault();
-                document.removeEventListener('mouseup', this.boundSwipeHandler);
-                document.removeEventListener('mousemove', this.boundSwipeHandler);
-                this.#swipeEnd();
+                if ( this.isSwiping ) {
+                    document.removeEventListener('mouseup', this.boundSwipeHandler);
+                    document.removeEventListener('mousemove', this.boundSwipeHandler);
+                    this.#swipeEnd();
+                }
                 break;
         }
     }
@@ -109,9 +111,9 @@ export default class DragInput{
     }
     #swipeMove(){
         const direction = this.#swipeDirection();
-        const swipeLength = Math.round( Math.abs(this.curX - this.prevX) ) + this.pixelsCorrection;
+        const swipeLength = Math.round( Math.abs(this.curX - this.prevX) * this.settings.dragModifier ) + this.pixelsCorrection;
 
-        if ( swipeLength <= this.threshold) return; // Ignore if less than 1 frame
+        if ( swipeLength <= this.threshold) return;// Ignore if less than 1 frame
         if ( direction !== 'left' && direction !== 'right') return; // Ignore vertical directions
 
         this.prevX = this.curX;
@@ -122,7 +124,7 @@ export default class DragInput{
         deltaFrames = deltaFrames % this.data.totalImages;
         // Add pixels to the next swipeMove if frames equivalent of swipe is not an integer number,
         // e.g one frame is 10px, swipeLength is 13px, we change 1 frame and add 3px to the next swipe,
-        // so fullwidth swipe is always rotate sprite for 1 turn
+        // so fullwidth swipe is always rotate sprite for 1 turn (with 'dragModifier' = 1)
         this.pixelsCorrection = swipeLength - (this.threshold * deltaFrames);
         this.changeFrame(this.getNextFrame( deltaFrames, (direction === 'left') )); // left means backward (reverse: true)
         this.data.canvas.element.dispatchEvent( new CustomEvent('animate-images:drag-change',